requirements.txt
# Core dependencies
python-dotenv>=1.0.0
structlog>=24.1.0
pyyaml>=6.0.1
pydantic>=2.6.0
telethon>=1.34.0  

# Platform-specific dependencies
# MetaTrader5 is Windows-only and requires MT5 terminal installation
# Install manually: pip install MetaTrader5==5.0.45
# Note: MT5 package is optional for development/testing on non-Windows platforms
MetaTrader5>=5.0.49; platform_system == "Windows" or (platform_system == "Darwin" and platform_machine == "x86_64")

# Data handling and analytics
pandas>=2.1.3
numpy>=1.26.2

# Testing and development
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0    # For mocking in tests
black>=23.11.0
mypy>=1.7.1
pylint>=3.0.2

# Logging and monitoring
structlog>=23.2.0

# Optional UI components
streamlit>=1.28.2  # For dashboard
fastapi>=0.104.1   # For API endpoints
uvicorn>=0.24.0    # ASGI server for FastAPI

# Development tools
pre-commit>=3.5.0  # For git hooks
isort>=5.12.0     # For import sorting
flake8>=6.1.0     # For additional linting

# Added from the code block
rich>=13.7.0
lark-parser>=0.12.0  # For robust signal parsing
alembic>=1.13.1
psycopg2-binary>=2.9.9  # Required for PostgreSQL
asyncpg>=0.29.0  # Required for async PostgreSQL operations 

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Telegram
python-telegram-bot>=20.6

# Trading
MetaTrader5>=5.0.45

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0
pyyaml>=6.0.1

# Development
black>=23.10.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Added from the code block
openai>=1.12.0